run npm lint and rename components to multi-word
