+++ front-end/src/components/HelloWorld.vue
@@ -2,8 +2,7 @@
   <div class="hello">
     <h1>{{ msg }}</h1>
     <p>
-      <a href="https://cli.vuejs.org" target="_blank" rel="noopener"
-        >vue-cli </a
+      <a href="https://cli.vuejs.org" target="_blank" rel="noopener">vue-cli </a
       >.
     </p>
     <h3>Installed CLI Plugins</h3>
