+++ front-end/src/helpers/api.js
@@ -1,9 +1,9 @@
 import axios from "axios";
 
-export const signupUser = (user) => {
-    return axios.post('/auth/register', user)
+export const signupUser = user => {
+  return axios.post("/auth/register", user);
 };
 
-export const signinUser = (user) => {
-    return  axios.post('/auth/login', user)
+export const signinUser = user => {
+  return axios.post("/auth/login", user);
 };
