+++ front-end/src/main.js
@@ -6,11 +6,11 @@ import i18n from "./i18n";
 import axios from "axios";
 
 Vue.config.productionTip = false;
-axios.defaults.baseURL = 'http://localhost:3000';
+axios.defaults.baseURL = "http://localhost:3000";
 
 new Vue({
-    router,
-    store,
-    i18n,
-    render: h => h(App)
+  router,
+  store,
+  i18n,
+  render: h => h(App)
 }).$mount("#app");
